Further fixes for web app redesign
* Import zepto/nav.js in built JS
* Fix SCSS lint errors
* Put footer into its own component
* Fix links in header
* Add `www_url` config option
