+++ components/builder-web/app/footer/_footer.scss
@@ -1,4 +1,4 @@
-.main-footer {
+.hab-footer {
   background-color: $hab-green-light;
   background-image: url("../assets/images/footer-mask.png"), linear-gradient(240deg, lighten($hab-green-light, 3%), $hab-green-light);
   background-position: center top;
