+++ components/builder-web/habitat/config/habitat.conf.js
@@ -6,4 +6,5 @@ habitatConfig({
     github_client_id: "{{cfg.github_client_id}}",
     source_code_url: "{{cfg.source_code_url}}",
     tutorials_url: "{{cfg.tutorials_url}}",
+    www_url: "{{cfg.www_url}}",
 });
