+++ components/builder-web/index.html
@@ -17,7 +17,5 @@
     </script>
     <script src="/habitat.conf.js"></script>
     <script id="hab-js" src="/assets/app.js"></script>
-    <script src="/assets/zepto.min.js"></script>
-    <script src="/assets/nav.js"></script>
   </body>
 </html>
